Move casper_propose_and_deploy into tool directory
