+++ integration-testing/test/test_star_network.py
@@ -1,5 +1,5 @@
 import pytest
-import test.casper_propose_and_deploy
+import tools.casper_propose_and_deploy
 from tools.network import start_network, wait_for_started_network, wait_for_converged_network
 from tools.rnode import start_bootstrap
 import logging
@@ -28,4 +28,4 @@ def test_convergence(star_network):
 
 @pytest.mark.skip(reason="This doesn't work since the show-blocks functionality was removed")
 def test_casper_propose_and_deploy(star_network):
-    test.casper_propose_and_deploy.run(star_network)
+    tools.casper_propose_and_deploy.run(star_network)
